changeFrame au bon endroit utilisation de geometry_utils si calcul de distance
