+++ common/geometry_utils/package.xml
@@ -46,6 +46,7 @@
   <build_depend>rospy</build_depend>
   <build_depend>std_msgs</build_depend>
   <build_depend>occupancy_grid_utils</build_depend>
+  <build_depend>tf</build_depend>
 
 
   <run_depend>geometry_msgs</run_depend>
@@ -54,6 +55,7 @@
   <run_depend>rospy</run_depend>
   <run_depend>std_msgs</run_depend>
   <run_depend>occupancy_grid_utils</run_depend>
+  <run_depend>tf</run_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
