+++ navigation/localisation/CMakeLists.txt
@@ -15,6 +15,7 @@ find_package(catkin REQUIRED COMPONENTS
   deplacement_msg
   tf
   message_filters
+	geometry_utils
 )
 
 ## System dependencies are found with CMake's conventions
@@ -88,7 +89,7 @@ find_package(Eigen REQUIRED)
 catkin_package(
    INCLUDE_DIRS include/localisation /usr/include/eigen3
    LIBRARIES localisation
-   CATKIN_DEPENDS geometry_msgs nav_msgs roscpp rospy sensor_msgs message_runtime
+   CATKIN_DEPENDS geometry_msgs nav_msgs roscpp rospy sensor_msgs message_runtime geometry_utils
 #  DEPENDS system_lib
 )
