+++ navigation/localisation/package.xml
@@ -50,6 +50,7 @@
   <build_depend>deplacement_msg</build_depend>
   <build_depend>tf</build_depend>
   <build_depend>message_filters</build_depend>
+  <build_depend>geometry_utils</build_depend>
 
   <run_depend>geometry_msgs</run_depend>
   <run_depend>nav_msgs</run_depend>
@@ -60,6 +61,7 @@
   <run_depend>deplacement_msg</run_depend>
   <run_depend>tf</run_depend>
   <run_depend>message_filters</run_depend>
+  <run_depend>geometry_utils</run_depend>
